+++ src/components/acter/landing-page/__stories__/info-seciton.stories.tsx
