+++ src/components/acter/landing-page/info-section/header.tsx
