+++ src/components/acter/landing-page/info-section/index.tsx
@@ -1,9 +1,9 @@
 import React, { FC } from 'react'
 import { makeStyles, Theme } from '@material-ui/core/styles'
 import { Card, CardContent } from '@material-ui/core'
-import Header from './info-section/header'
-import PeopleSection from './info-section/people'
-import Organzations from './info-section/organizations'
+import Header from 'src/components/acter/landing-page/info-section/header'
+import PeopleSection from 'src/components/acter/landing-page/info-section/people'
+import Organzations from 'src/components/acter/landing-page/info-section/organizations'
 
 import { Acter } from '@generated/type-graphql'
