+++ src/components/acter/landing-page/info-section/organizations.tsx
