+++ src/components/acter/landing-page/info-section/people.tsx
